Updates to structure
Dealing with new pathing with by/fy separation
